Move TestSuiteBase plugin to testing-base
