+++ subprojects/core/src/integTest/groovy/org/gradle/api/internal/plugins/ApplyPluginBuildOperationIntegrationTest.groovy
@@ -49,7 +49,7 @@ class ApplyPluginBuildOperationIntegrationTest extends AbstractIntegrationSpec {
             "org.gradle.api.plugins.BasePlugin": null,
             "org.gradle.language.base.plugins.LifecycleBasePlugin": null,
             "org.gradle.api.plugins.ReportingBasePlugin": null,
-            "org.gradle.platform.base.plugins.TestSuiteBasePlugin": "org.gradle.test-suite-base",
+            "org.gradle.testing.base.plugins.TestSuiteBasePlugin": "org.gradle.test-suite-base",
             "org.gradle.api.plugins.JvmTestSuitePlugin": "org.gradle.jvm-test-suite",
         ]
