+++ subprojects/testing-base/src/main/resources/META-INF/gradle-plugins/org.gradle.test-suite-base.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=org.gradle.platform.base.plugins.TestSuiteBasePlugin
+implementation-class=org.gradle.testing.base.plugins.TestSuiteBasePlugin
